quality-gate/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile requirements.in
#
certifi==2026.1.4
    # via requests
cffi==2.0.0
    # via cryptography
charset-normalizer==3.4.4
    # via requests
cryptography==46.0.5
    # via google-auth
google-auth==2.48.0
    # via
    #   -r requirements.in
    #   google-auth-oauthlib
    #   gspread
google-auth-oauthlib==1.2.4
    # via gspread
gspread==6.2.1
    # via -r requirements.in
idna==3.11
    # via requests
oauthlib==3.3.1
    # via requests-oauthlib
pyasn1==0.6.2
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pycparser==3.0
    # via cffi
python-dotenv==1.2.1
    # via -r requirements.in
requests==2.32.5
    # via
    #   -r requirements.in
    #   requests-oauthlib
requests-oauthlib==2.0.0
    # via google-auth-oauthlib
rsa==4.9.1
    # via google-auth
urllib3==2.6.3
    # via requests
